refactor: inline useCamera into useCanvasGestures and delete it
useCamera created local useState, making it look like a reusable hook
when it was really single-use. Inlining it into its only consumer
prevents anyone from importing it and silently getting independent
camera state.
